Address fbalak review comments on CephFS orphaned snapshot alert
- Move find_consumer_for_storage_client and get_consumer_svg_on_provider
  from helpers/odf_cephfs_snap.py to ocs/resources/storageconsumer.py,
  where generic StorageConsumer helpers belong
- Add inline comment clarifying that consumer name equals SVG name
  (current mapping; may change in future versions)
- Replace log.info("Step X: ...") calls with log.test_step() throughout
  the test for proper step tracking; remove redundant step-number prefixes
  since TEST_STEP log level auto-numbers steps
