+++ ocs_ci/helpers/odf_cephfs_snap.py
@@ -5,8 +5,8 @@ from ocs_ci.framework import config
 from ocs_ci.helpers.helpers import get_snapshot_content_obj
 from ocs_ci.ocs import constants, ocp
 from ocs_ci.utility import templating
-from ocs_ci.ocs.managedservice import get_consumer_names
 from ocs_ci.ocs.resources import ocs
+from ocs_ci.ocs.resources.storageconsumer import find_consumer_for_storage_client
 from ocs_ci.utility.utils import TimeoutSampler
 
 log = logging.getLogger(__name__)
@@ -91,7 +91,7 @@ def create_provider_retain_cephfs_snapclass(snapclass_name, storage_client_name)
         ), f"Failed to create VolumeSnapshotClass {snapclass_name} on provider"
         log.info("Created VolumeSnapshotClass %s on provider", snapclass_name)
 
-        consumer_name, consumer_data = _find_consumer_for_storage_client(
+        consumer_name, consumer_data = find_consumer_for_storage_client(
             storage_client_name
         )
         consumer_ocp = ocp.OCP(
@@ -147,58 +147,6 @@ def create_provider_retain_cephfs_snapclass(snapclass_name, storage_client_name)
             log.info("Deleted VolumeSnapshotClass %s from provider", snapclass_name)
 
     return _teardown
-
-
-def _find_consumer_for_storage_client(storage_client_name):
-    """
-    Find the StorageConsumer CR on the provider that owns ``storage_client_name``.
-
-    Must be called within an active provider config context.
-
-    Args:
-        storage_client_name (str): StorageClient CR name on the client cluster.
-
-    Returns:
-        tuple[str, dict]: ``(consumer_name, consumer_data)`` for the matching
-            StorageConsumer.
-
-    Raises:
-        AssertionError: If no matching StorageConsumer is found.
-    """
-    consumer_ns = config.ENV_DATA["cluster_namespace"]
-    consumer_ocp = ocp.OCP(kind=constants.STORAGECONSUMER, namespace=consumer_ns)
-    for name in get_consumer_names():
-        consumer_data = consumer_ocp.get(resource_name=name)
-        if (
-            consumer_data.get("status", {}).get("client", {}).get("name")
-            == storage_client_name
-        ):
-            return name, consumer_data
-    raise AssertionError(
-        f"No StorageConsumer found for storage client '{storage_client_name}'"
-    )
-
-
-def get_consumer_svg_on_provider(storage_client_name):
-    """
-    Return the CephFS subvolume group name for a storage client on the
-    provider cluster.
-
-    The SVG name equals the StorageConsumer CR name on the provider.
-    Must be called within an active provider config context.
-
-    Args:
-        storage_client_name (str): StorageClient CR name on the client
-            cluster.
-
-    Returns:
-        str: Subvolume group name on the provider cluster.
-
-    Raises:
-        AssertionError: If no matching StorageConsumer is found.
-    """
-    consumer_name, _ = _find_consumer_for_storage_client(storage_client_name)
-    return consumer_name
 
 
 def get_cephfs_snap_entries(snap_runner):
